requirements.txt
discord.py
pyaml
requests
